feat: 결제수단 스키마 세팅
