+++ package.json
@@ -18,6 +18,7 @@
     "swagger-typescript-api": "swagger-typescript-api generate -p http://15.165.214.75:8080/v3/api-docs -r -o ./apis --modular -d --extract-request-body --extract-response-body --extract-response-error --axios --clean-output"
   },
   "dependencies": {
+    "@hookform/resolvers": "^5.2.2",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-select": "^2.2.6",
@@ -31,7 +32,8 @@
     "react-dom": "^19.1.1",
     "react-router-dom": "^7.9.6",
     "tailwind-merge": "^3.3.1",
-    "vite-tsconfig-paths": "^5.1.4"
+    "vite-tsconfig-paths": "^5.1.4",
+    "zod": "^4.1.12"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^4.1.2",
