+++ src/router/constant/routes.ts
@@ -1,3 +1,4 @@
 export const ROUTES = {
   HOME: '/',
+  PAYMENT: '/payment'
 };
