+++ src/router/router.tsx
@@ -3,6 +3,7 @@ import { createBrowserRouter } from 'react-router-dom';
 import { ROUTES } from '@router/constant/routes';
 import Layout from '@router/Layout';
 import Home from '@/pages/home/Home';
+import PaymentInfo from '@/pages/pay/PaymentInfo'
 
 export const router = createBrowserRouter([
   {
@@ -11,6 +12,10 @@ export const router = createBrowserRouter([
       {
         path: ROUTES.HOME,
         element: <Home />,
+      },
+      {
+        path: ROUTES.PAYMENT,
+        element: <PaymentInfo />,
       },
     ],
   },
